+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Bug #2: Update assertion messages for `Enum` class tests (@terabytesoftw)
 - Enh #3: Introduce `Encode` helper class with tests (@terabytesoftw)
 - Bug #4: Add test cases to kill bitwise operator mutations in `Encode::value()` (@terabytesoftw)
+- Enh #5: Introduce `Attribute` helper class with tests (@terabytesoftw)